Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
     float pi=3.14;
-    int a;
-    scanf("%d", &a);
+    float a;
+    scanf("%f", &a);
     float area=pi*a*a;
     printf("Area: %f",area);
 }
